refactor: reorganizada estrutura do projeto agrupando arquivos por funcionalidade
