+++ front/src/renderer/src/components/MenuLateral/MenuLateral.css
